refactor: implement useControls for managing view updates in Controls component
